feat: Implement #799
[NITPICK] automation/__init__.py inconsistent re-export vs validation/__init__.py

Closes #799
