+++ tests/unit/automation/test_package_imports.py
@@ -1,11 +1,37 @@
-"""Regression tests for automation package import side effects."""
+"""Regression tests for automation package import side effects and public API contract.
+
+This module tests both import-side-effect safety and the public re-export surface.
+Issue #799: the package must re-export every reviewer class and option model at
+the root, so consumers do not need to reach into private-looking submodules.
+"""
 
 from __future__ import annotations
 
+import importlib
 import subprocess
 import sys
 
 import pytest
+
+import hephaestus.automation as automation
+
+EXPECTED_PUBLIC_SYMBOLS = {
+    "AddressReviewOptions": "hephaestus.automation.models",
+    "AddressReviewer": "hephaestus.automation.address_review",
+    "AuditReviewer": "hephaestus.automation.audit_reviewer",
+    "CIDriver": "hephaestus.automation.ci_driver",
+    "CIDriverOptions": "hephaestus.automation.models",
+    "DependencyResolver": "hephaestus.automation.dependency_resolver",
+    "ImplementerOptions": "hephaestus.automation.models",
+    "IssueImplementer": "hephaestus.automation.implementer",
+    "IssueInfo": "hephaestus.automation.models",
+    "PRReviewer": "hephaestus.automation.pr_reviewer",
+    "PlanReviewer": "hephaestus.automation.plan_reviewer",
+    "PlanReviewerOptions": "hephaestus.automation.models",
+    "Planner": "hephaestus.automation.planner",
+    "PlannerOptions": "hephaestus.automation.models",
+    "ReviewerOptions": "hephaestus.automation.models",
+}
 
 _PHASE_ENTRYPOINTS = (
     "hephaestus.automation.address_review",
@@ -42,18 +68,32 @@ if loaded:
     assert result.returncode == 0, result.stderr + result.stdout
 
 
-def test_package_lazy_exports_remain_importable() -> None:
-    """Existing package-level class imports should keep working."""
-    result = _run_python(
-        """
-import hephaestus.automation as automation
+def test_all_matches_expected() -> None:
+    """Assert __all__ exactly matches the expected public API."""
+    assert set(automation.__all__) == set(EXPECTED_PUBLIC_SYMBOLS)
 
-for name in automation.__all__:
-    assert getattr(automation, name).__name__ == name
-"""
-    )
 
-    assert result.returncode == 0, result.stderr + result.stdout
+def test_every_all_entry_resolves_via_package_root() -> None:
+    """Assert every export is accessible and points to the correct source module.
+
+    This test is stronger than a simple __name__ check: it verifies that the
+    re-exported object is the same object (same identity) as in the source module.
+    This ensures there are no accidental copies or rewraps of public classes.
+    """
+    for name, source_module in EXPECTED_PUBLIC_SYMBOLS.items():
+        value = getattr(automation, name)
+        source = importlib.import_module(source_module)
+        assert value is getattr(source, name), (
+            f"hephaestus.automation.{name} is not the same object as "
+            f"{source_module}.{name}"
+        )
+
+
+def test_dir_includes_all_entries() -> None:
+    """Assert dir() introspection includes all public exports."""
+    listing = dir(automation)
+    for name in EXPECTED_PUBLIC_SYMBOLS:
+        assert name in listing, f"{name!r} missing from dir(hephaestus.automation)"
 
 
 def test_public_surface_pins_expected_symbols() -> None:
